+++ src/compare.c
@@ -1,11 +1,11 @@
 /*-
  ***********************************************************************
  *
- * $Id: compare.c,v 1.52 2013/02/14 16:55:19 mavrik Exp $
+ * $Id: compare.c,v 1.54 2014/07/30 07:07:30 mavrik Exp $
  *
  ***********************************************************************
  *
- * Copyright 2000-2013 The FTimes Project, All Rights Reserved.
+ * Copyright 2000-2014 The FTimes Project, All Rights Reserved.
  *
  ***********************************************************************
  */
@@ -591,7 +591,7 @@ CompareLoadBaselineData(SNAPSHOT_CONTEXT *psBaseline, char *pcError)
       snprintf(pcError, MESSAGE_SIZE, "%s: CreateFile(): %s", acRoutine, pcMessage);
       return ER;
     }
-    iFile = _open_osfhandle((long) hFile, 0);
+    iFile = _open_osfhandle((intptr_t) hFile, 0);
     if (iFile == ER)
     {
       snprintf(pcError, MESSAGE_SIZE, "%s: open_osfhandle(): Handle association failed.", acRoutine);
